add: api for product

+++ FridgeApp.Application/Commands/CreateProduct.cs
@@ -3,5 +3,5 @@ using FridgeApp.Shared.Abstractions.Commands;
 
 namespace FridgeApp.Application.Commands
 {
-    public record CreateProduct(Guid Id, string Name, ushort DefaultQuantity) : ICommand;
+    public record CreateProduct(Guid Id, string Name, int DefaultQuantity) : ICommand;
 }
