+++ FridgeApp.Application/DTOs/FridgeModelDto.cs
@@ -5,7 +5,7 @@ namespace FridgeApp.Application.DTOs
     public class FridgeModelDto
     {
         public string Name { get; set; }
-        public ushort Year { get; set; }
+        public int Year { get; set; }
         public int Version { get; set; }
         public DateTimeOffset CreatedDateTime { get; set; }
         public DateTimeOffset? UpdatedDateTime { get; set; }
