+++ FridgeApp.Application/DTOs/FridgeProductDto.cs
@@ -6,8 +6,8 @@ namespace FridgeApp.Application.DTOs
     {
         public Guid Id { get; set; }
         public string Name { get; set; }
-        public ushort DefaultQuantity { get; set; }
-        public ushort Quantity { get; set; }
+        public int DefaultQuantity { get; set; }
+        public int Quantity { get; set; }
         public int Version { get; set; }
         public DateTimeOffset CreatedDateTime { get; set; }
         public DateTimeOffset? UpdatedDateTime { get; set; }
