+++ FridgeApp.Domain/Events/FridgeProductAdded.cs
@@ -3,5 +3,5 @@ using FridgeApp.Shared.Abstractions.Domain;
 
 namespace FridgeApp.Domain.Events
 {
-    public record FridgeProductAdded(Fridge Fridge, Product Product, ushort Quantity) : IDomainEvent;
+    public record FridgeProductAdded(Fridge Fridge, Product Product, int Quantity) : IDomainEvent;
 }
